backend/requirements.txt
# FastAPI Backend
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# OpenAI (for embeddings and ChatKit/Agents)
openai==1.54.0

# Qdrant Vector Store
qdrant-client==1.12.0

# Database (Neon Postgres)
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
tenacity==8.2.3  # Retry logic for robust API calls

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
